+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [Unreleased]
+### Added
+- Added auto-registration for the doctrine dbal enum types.
+- Added optional configuration parameter `source_directories`.
+
 ## [1.0.0] - 2019-05-17
 ### Added
 - First release of this bundle.